# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://github.com/EDDYCJY/go-gin-example/blob/master/LICENSE
  termsOfService: https://github.com/EDDYCJY/go-gin-example
  title: Golang Gin API
  version: "1.0"
paths:
  /api/v1/tags/import:
    post:
      parameters:
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Import Image
  /auth:
    get:
      parameters:
      - description: userName
        in: query
        name: username
        required: true
        type: string
      - description: password
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get Auth
  /starling:
    delete:
      consumes:
      - application/json
      description: 删除镜像迁移任务
      parameters:
      - description: ids
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 删除镜像迁移任务
    get:
      consumes:
      - application/json
      description: 根据字段分页返回
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: page_size
        in: path
        name: page_size
        required: true
        type: string
      - description: page_num
        in: path
        name: page_num
        required: true
        type: string
      - description: chinese_text
        in: path
        name: chinese_text
        required: true
        type: string
      - description: english_text
        in: path
        name: english_text
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: 分页搜索
    post:
      parameters:
      - description: Key
        in: query
        name: key
        required: true
        type: string
      - description: chinese_text
        in: query
        name: ChineseText
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Post Starling
    put:
      parameters:
      - description: id
        in: query
        name: ID
        required: true
        type: string
      - description: key
        in: query
        name: key
        required: true
        type: string
      - description: chinese_text
        in: query
        name: ChineseText
        required: true
        type: string
      - description: english_text
        in: query
        name: englishText
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Post Starling
  /starling/batch:
    post:
      parameters:
      - description: key
        in: query
        name: key
        required: true
        type: string
      - description: chinese_text
        in: query
        name: ChineseText
        required: true
        type: string
      - description: english_text
        in: query
        name: EnglishText
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Post Starling
swagger: "2.0"